+++ Pandora/Paradigm/Structure/Ability/Accessible.hs
@@ -1,0 +1,6 @@
+module Pandora.Paradigm.Structure.Ability.Accessible where
+
+import Pandora.Paradigm.Inventory.Optics (type (:-.))
+
+class Accessible a e where
+	access :: a :-. e